move advanced integrations up one page
